+++ src/command/scrub/setup.rs
@@ -10,9 +10,9 @@ use crate::{
         scrub::{
             check::{
                 BlobChecker, BlobIndexChecker, DigestLinkChecker, LinkReferencesChecker,
-                ManifestChecker, MediaTypeChecker, MultipartChecker, NamespaceChecker,
-                OrphanGrantChecker, OrphanJobChecker, OrphanNamespaceChecker, OrphanQueue,
-                ReferrerChecker, StoreChecker, TagChecker, UploadChecker,
+                ManifestChecker, MultipartChecker, NamespaceChecker, OrphanGrantChecker,
+                OrphanJobChecker, OrphanNamespaceChecker, OrphanQueue, ReferrerChecker,
+                StoreChecker, TagChecker, UploadChecker,
             },
             command::Options,
             error::Error,
@@ -77,13 +77,6 @@ pub fn namespace_checkers(
 
     if options.reconcile_blob_index {
         checkers.push(Box::new(BlobIndexChecker::new(
-            blob_store.clone(),
-            metadata_store.clone(),
-        )));
-    }
-
-    if options.media_types {
-        checkers.push(Box::new(MediaTypeChecker::new(
             blob_store.clone(),
             metadata_store.clone(),
         )));